refactor, reset filters calculated depending on the filters applied
