+++ src/Components/Common/Tables.js
@@ -7,6 +7,10 @@ import {
   FILTER_CATEGORIES,
   RULE_CATEGORIES,
 } from '../../AppConstants';
+import {
+  hasAnyValueGreaterThanZero,
+  remappingSeverity,
+} from '../../Utilities/Workloads';
 
 export const passFilters = (rule, filters) =>
   Object.entries(filters).every(([filterKey, filterValue]) => {
@@ -275,48 +279,6 @@ export const addFilterParam = (currentFilters, updateFilters, param, values) =>
         ...{ [param]: values },
       })
     : removeFilterParam(currentFilters, updateFilters, param);
-
-export const severityTypeToText = (value) => {
-  value = parseInt(value);
-  if (value === 1) {
-    return 'Low';
-  } else if (value === 2) {
-    return 'Moderate';
-  } else if (value === 3) {
-    return 'Important';
-  } else {
-    return 'Critical';
-  }
-};
-
-export const remappingSeverity = (obj, mode) => {
-  const mapping = {
-    1: 'low',
-    2: 'moderate',
-    3: 'important',
-    4: 'critical',
-  };
-  let updatedObj = {};
-
-  if (mode === 'general') {
-    for (const key in obj) {
-      if (key in mapping) {
-        updatedObj[mapping[key]] = obj[key];
-      }
-    }
-  } else {
-    updatedObj = mapping[obj];
-  }
-  return updatedObj;
-};
-
-function hasAnyValueGreaterThanZero(obj, stringsToCheck) {
-  for (const key of stringsToCheck) {
-    if (obj[key] > 0) {
-      return true; // Return true if any matching string has a value greater than 0
-    }
-  }
-}
 
 export const passFilterWorkloads = (workloads, filters) => {
   const generalSeverityRemapped = remappingSeverity(
