+++ src/Components/HighestSeverityBadge/HighestSeverityBadge.js
@@ -3,7 +3,7 @@ import { Tooltip } from '@patternfly/react-core';
 import { TooltipPosition } from '@patternfly/react-core/dist/js/components/Tooltip';
 import InsightsLabel from '@redhat-cloud-services/frontend-components/InsightsLabel';
 import PropTypes from 'prop-types';
-import { severityTypeToText } from '../Common/Tables';
+import { severityTypeToText } from '../../Utilities/Workloads';
 
 export const HighestSeverityBadge = ({ highestSeverity, severities }) => {
   const severitiesToDisplay = Object.keys(severities)
