+++ src/Utilities/Workloads.js
@@ -1,3 +1,5 @@
+import _, { isEmpty } from 'lodash';
+
 export const SEVERITY_OPTIONS = [
   {
     value: 'critical',
@@ -33,7 +35,7 @@ export const SEVERITY_OPTIONS = [
   },
 ];
 
-export const remappingSeverity = (obj) => {
+export const remappingSeverity = (obj, mode) => {
   const mapping = {
     1: 'low',
     2: 'moderate',
@@ -42,10 +44,45 @@ export const remappingSeverity = (obj) => {
   };
   let updatedObj = {};
 
-  for (const key in obj) {
-    if (key in mapping) {
-      updatedObj[mapping[key]] = obj[key];
+  if (mode === 'general' || mode === 'label') {
+    for (const key in obj) {
+      if (key in mapping) {
+        updatedObj[mapping[key]] = obj[key];
+      }
+    }
+  } else {
+    updatedObj = mapping[obj];
+  }
+
+  return updatedObj;
+};
+
+export const hasAnyValueGreaterThanZero = (obj, stringsToCheck) => {
+  for (const key of stringsToCheck) {
+    if (obj[key] > 0) {
+      return true; // Return true if any matching string has a value greater than 0
     }
   }
-  return updatedObj;
+};
+
+export const severityTypeToText = (value) => {
+  value = parseInt(value);
+  if (value === 1) {
+    return 'Low';
+  } else if (value === 2) {
+    return 'Moderate';
+  } else if (value === 3) {
+    return 'Important';
+  } else {
+    return 'Critical';
+  }
+};
+
+export const noFiltersApplied = (params) => {
+  const cleanedUpParams = _.cloneDeep(params);
+  delete cleanedUpParams.sortIndex;
+  delete cleanedUpParams.sortDirection;
+  delete cleanedUpParams.offset;
+  delete cleanedUpParams.limit;
+  return Object.values(cleanedUpParams).filter((value) => !isEmpty(value));
 };
